pyproject.toml
[tool.poetry]
name = "snake"
version = "0.1.0"
description = ""
authors = ["Carlos Frey <carlosfrey@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
pynput = "^1.7.7"
websockets = "^14.1"
json-numpy = "^2.1.0"
watchpoints = "^0.2.5"
asyncio = "^3.4.3"
pydantic = "^2.10.3"

[tool.poetry.scripts]
start = 'snake.scripts.launcher:start'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
